Make a hack for LTO work only when you are actually doing LTO.
We allow an archive file without symbol table as a linker input as a
workaround for a very common error in LTO build. But that logic worked
even for an archive file containing non-bitcode files, which is not
expected. This patch limits that workaround to one that contains only
bitcode files.

Differential Revision: https://reviews.llvm.org/D59373

git-svn-id: http://llvm.org/svn/llvm-project/lld/trunk@356186 91177308-0d34-0410-b5e6-96231b3b80d8