[LLD][COFF] Implement /filealign parameter
Patch by Stefan Schmidt.

This adds the /filealign parameter to lld, which allows to specify the
section alignment in the output file (as it does on Microsoft's
link.exe).

This is required to be able to load dynamically linked libraries on the
original Xbox, where the debugger monitor expects the section alignment
in the file to be the same as in memory.

git-svn-id: http://llvm.org/svn/llvm-project/lld/trunk@361634 91177308-0d34-0410-b5e6-96231b3b80d8